Improved level filtering (#43)
* Create a SongSelectIndex
The current song select uses a MapIndex and its id for indexing, this aims to remove the map id as a key and allow duplicate entryes to share the same map and map id.

* Refactor song selection and filters to use SongSelectIndex
Since these changes involved adding and using a new type, I didn't want to split the commits to leave versions that didn't compile.
This changes uses SongSelectIndex instead of MapIndex*, and keys on the id of the SongSelectIndex instead as well. Filters have been updated with these changes as well, and by default the GUI supports the correct visualization of filtered difficulties simply by using the new system.

* Fix an out-of-range bug by referencing the correct collection

* Implement folder filter

* Remove problematic call
It looks like all code paths where this is called also invoke it again later after filtering has been reset, so this is redundant. The whole thing needs cleaned up..